Handle some functions with return type.
From a previous commit:

A template function has 'v' as argument, followed by the function
arguments, so we need to skip it.

This is complete wrong. The 'v' here is the return type of the function.
Some functions have the return type included (namely template
functions). This commit fixes it.

There are other functions with return types included which need to be
handled correctly.